refactor: consolidate transcription schemas and shared ops
Made-with: Cursor
