+++ packages/sdk/schemas/index.ts
@@ -19,8 +19,7 @@ export * from "./transcription";
 export * from "./translate";
 export * from "./translation-config";
 export * from "./llamacpp-config";
-export * from "./whispercpp-config";
-export * from "./parakeet-config";
+export * from "./transcription-config";
 export * from "./text-to-speech";
 export * from "./error";
 export * from "./rag";
