+++ packages/sdk/server/bare/plugins/whispercpp-transcription/plugin.ts
@@ -4,16 +4,18 @@ import TranscriptionWhispercpp, {
 } from "@qvac/transcription-whispercpp";
 import {
   definePlugin,
+  defineHandler,
+  transcribeStreamRequestSchema,
+  transcribeStreamResponseSchema,
   ModelType,
   type CreateModelParams,
   type PluginModelResult,
   type WhisperConfig,
 } from "@/schemas";
-import { createTranscribeStreamHandler } from "@/server/bare/utils/transcription-handler";
 import { ADDON_NAMESPACES, createStreamLogger } from "@/logging";
 import { parseModelPath } from "@/server/utils";
 import FilesystemDL from "@qvac/dl-filesystem";
-import { transcribe } from "@/server/bare/plugins/whispercpp-transcription/ops/transcribe-stream";
+import { transcribe } from "@/server/bare/ops/transcribe";
 
 function createWhisperModel(
   modelId: string,
@@ -77,7 +79,30 @@ export const whisperPlugin = definePlugin({
   },
 
   handlers: {
-    transcribeStream: createTranscribeStreamHandler(transcribe),
+    transcribeStream: defineHandler({
+      requestSchema: transcribeStreamRequestSchema,
+      responseSchema: transcribeStreamResponseSchema,
+      streaming: true,
+
+      handler: async function* (request) {
+        for await (const text of transcribe({
+          modelId: request.modelId,
+          audioChunk: request.audioChunk,
+          prompt: request.prompt,
+        })) {
+          yield {
+            type: "transcribeStream" as const,
+            text,
+          };
+        }
+
+        yield {
+          type: "transcribeStream" as const,
+          text: "",
+          done: true,
+        };
+      },
+    }),
   },
 
   logging: {
